+++ .mocharc.js
@@ -1,0 +1,5 @@
+module.exports = {
+  require: 'ts-node/register',
+  extensions: ['ts'],
+  spec: ['test/**/*.spec.*'],
+};