+++ babel.config.js
@@ -1,6 +1,0 @@
-module.exports = function (api) {
-  api.cache.never();
-  return {
-    presets: [['@babel/preset-env', { targets: { node: 'current' } }]],
-  };
-};