+++ cli/csv-to-xlsx.ts
@@ -1,0 +1,78 @@
+#!/usr/bin/env node
+
+import path from 'path';
+import fs from 'fs-extra';
+import { program } from 'commander';
+
+import convertCsvToXlsx from '../src/convertCsvToXlsx';
+import pkg from '../package.json';
+
+program
+  .version(pkg.version, '-v, --version')
+  .option(
+    '-i, --input-dir [dir]',
+    'Input directory that has the CSV files',
+    'csv',
+  )
+  .option(
+    '-o, --output-dir [dir]',
+    'Output directory for the XLSX files',
+    'xlsx',
+  )
+  .option(
+    '-s, --sheet-name [name]',
+    'Set the sheet name to be used in the XLSX files',
+    '',
+  );
+
+program.on('--help', function () {
+  console.log(``);
+  console.log(`Created by: ${pkg.author.name}`);
+  console.log(`Please report bugs at: ${pkg.bugs.url}`);
+  console.log(`Version: ${pkg.version}`);
+});
+
+program.parse(process.argv);
+
+const programOptions = program.opts();
+
+const csvPath = path.join(process.cwd(), programOptions.inputDir);
+const xlsxPath = path.join(process.cwd(), programOptions.outputDir);
+
+// check the csvPath
+if (!fs.existsSync(csvPath)) {
+  // csv folder doesn't exist, doing it wrong
+  console.error(`Invalid input directory: ${csvPath}\n`);
+  process.exitCode = 1;
+  program.help(); // exit immediately
+}
+
+// check the xlsxPath
+if (!fs.existsSync(xlsxPath)) {
+  // create xlsx folder
+  console.info(`Creating output directory: ${xlsxPath}`);
+  fs.mkdirSync(xlsxPath, { recursive: true });
+}
+
+// read csvPath
+const csvFiles = fs.readdirSync(csvPath);
+
+for (const file of csvFiles) {
+  // parse file
+  const fileObject = path.parse(file);
+  // check file extension
+  if (fileObject.ext !== '.csv') {
+    continue;
+  }
+  console.info(`Converting: ${fileObject.name}`);
+  // convert
+  try {
+    const source = path.join(csvPath, file);
+    const destination = path.join(xlsxPath, `${fileObject.name}.xlsx`);
+    convertCsvToXlsx(source, destination);
+  } catch (e) {
+    console.error(e.toString());
+  }
+}
+
+console.info(`Complete.`);